requirements.txt
openai
python-dotenv
gradio
Pillow
torch
numpy
safetensors
